% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_data.R
\name{grab_data}
\alias{grab_data}
\title{Load a drive file into a dataframe}
\usage{
grab_data(url, filetype, ...)
}
\arguments{
\item{url}{The link to the desired file obtained by "get shareable link" (the
user must have at least viewing access to the file). Can be either a `.csv`,
Google Sheets, or `.Rda` file.}

\item{filetype}{A character vector specifying the filetype. Either "csv",
"Sheet", or "Rda".}

\item{...}{Additional arguments to supply to `read_csv()`.}
}
\description{
This function grabs a dataset from Google Drive and returns
it as an R dataframe object. You might need to run `drive_auth()` before
running this to get your Google account credentials set up.
}
\examples{
\dontrun{
# Grab a very useful dataset
grab_data(url = "https://drive.google.com/open?id=1JnkQddF8FmY0YsXZSWBanQNPjZ_LBIQUGi_s43vNDIQ",
          filetype = "Sheet")

# Send extra arguments to read_csv
grab_data("https://drive.google.com/file/d/10j5nM5Wu27S5-kIsf1x6HlkLpOKvcqhE/view?usp=sharing",
          "csv",
          col_types = cols(compelling_narrative = col_character(),
                           index = col_integer()))
}
}
